requirements.txt
coreapi==2.3.3
Django==1.11.7
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.3
psycopg2==2.7.3.2
